refactor: New HeaderViewModel
